requirements.txt
discord.py
girc>=0.3.1
python-slugify
